TreeItemFoundAfterExpanding is now used in DefaultTreeItem - should be more stable when tree item is dynamically generated
