+++ org.jboss.reddeer.swt/src/org/jboss/reddeer/swt/impl/tree/AbstractTreeItem.java
@@ -14,7 +14,6 @@ import org.eclipse.swtbot.swt.finder.exceptions.WidgetNotFoundException;
 import org.eclipse.swtbot.swt.finder.widgets.SWTBotTree;
 import org.eclipse.swtbot.swt.finder.widgets.SWTBotTreeItem;
 import org.jboss.reddeer.swt.api.TreeItem;
-import org.jboss.reddeer.swt.condition.WaitCondition;
 import org.jboss.reddeer.swt.exception.SWTLayerException;
 import org.jboss.reddeer.swt.lookup.impl.WidgetLookup;
 import org.jboss.reddeer.swt.wait.WaitUntil;
@@ -242,74 +241,3 @@ public abstract class AbstractTreeItem implements TreeItem {
 	}
 	
 }
-
-/**
- * Condition is fulfilled when tree item node is found after expanding 
- * 
- * @author jjankovi
- *
- */
-class TreeItemFoundAfterExpanding implements WaitCondition {
-
-	private String treeItemNode;
-	private SWTBotTreeItem item;
-	
-	public TreeItemFoundAfterExpanding(SWTBotTreeItem item, String treeItemNode) {
-		super();
-		this.treeItemNode = treeItemNode;
-		this.item = item;
-	}
-	
-	@Override
-	public boolean test() {
-		item.expand();
-		return nodeIsFound(treeItemNode);		
-	}
-
-	@Override
-	public String description() {
-		return "Tree item '" + treeItemNode + "' not found.";
-	}
-	
-	private boolean nodeIsFound(String treeItemNode) {
-		
-		try {
-			item.getNode(treeItemNode);
-			System.out.println(treeItemNode + " was found");
-			return true;
-		} catch (WidgetNotFoundException wnfe) {
-			item.collapse();
-			return false;
-		}
-		
-	}
-	
-}
-
-///**
-// * Condition is fulfilled when tree item is selected
-// * 
-// * @author jjankovi
-// *
-// */
-//class TreeItemIsSelected implements WaitCondition {
-//
-//	private SWTBotTreeItem item; 
-//	
-//	public TreeItemIsSelected(SWTBotTreeItem item) {
-//		super();
-//		this.item = item;
-//	}
-//
-//	@Override
-//	public boolean test() {
-//		item.select();
-//		return item.isSelected();
-//	}
-//
-//	@Override
-//	public String description() {
-//		return "Tree item '" + item.getText() + "' cannot be selected.";
-//	}
-//	
-//}
